chore(notification): remove next and v11 toggle
