+++ packages/react/src/components/Notification/index.js
@@ -11,47 +11,32 @@ import {
   ToastNotification as ToastNotificationNext,
   InlineNotification as InlineNotificationNext,
   ActionableNotification as ActionableNotificationNext,
-} from './next/Notification';
-import {
-  NotificationActionButton as NotificationActionButtonClassic,
-  NotificationTextDetails as NotificationTextDetailsClassic,
-  NotificationButton as NotificationButtonClassic,
-  ToastNotification as ToastNotificationClassic,
-  InlineNotification as InlineNotificationClassic,
 } from './Notification';
 import { createComponentToggle } from '../../internal/ComponentToggle';
 
 export const NotificationActionButton = createComponentToggle({
   name: 'NotificationActionButton',
   next: NotificationActionButtonNext,
-  classic: NotificationActionButtonClassic,
-});
-
-export const NotificationTextDetails = createComponentToggle({
-  name: 'NotificationTextDetails',
-  classic: NotificationTextDetailsClassic,
 });
 
 export const NotificationButton = createComponentToggle({
   name: 'NotificationButton',
   next: NotificationButtonNext,
-  classic: NotificationButtonClassic,
 });
 
 export const ToastNotification = createComponentToggle({
   name: 'ToastNotification',
   next: ToastNotificationNext,
-  classic: ToastNotificationClassic,
 });
 
 export const InlineNotification = createComponentToggle({
   name: 'InlineNotification',
   next: InlineNotificationNext,
-  classic: InlineNotificationClassic,
 });
 
 export const ActionableNotification = createComponentToggle({
   name: 'ActionableNotification',
   next: ActionableNotificationNext,
-  classic: null,
 });
+
+// export * from './Notification';
